Possible bug fix for memory region enumeration
Not familiar with the code base at all, but I was reading through this file and it seems as though there is a possible typo? I'm not sure if the original behaviour was intended or not but I thought it was worth pointing this out in case it was an accident!